requirements.txt
pydantic==2.7.4
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38
langchain-text-splitters==0.2.2
langchain-community==0.2.16
faiss-cpu==1.13.2
PyPDF2
flask